Map Data: Write maximum number of ways in a tile as a short instead of a byte into the map data

+++ Osm2GpsMid/src/de/ueller/osmToGpsMid/CreateGpsMidData.java
@@ -1069,7 +1069,7 @@ public class CreateGpsMidData {
 			writeNode(n,ds,SEGNODE,t);
 		}
 		ds.writeByte(0x55); // Magic number
-		ds.writeByte(ways.size());
+		ds.writeShort(ways.size());
 		for (Way w : ways){
 			w.write(ds, names1,t);
 		}
